Feat: Get location, Naver map api
현 위치 받아, 네이버 맵에 아이콘 표시, 반경 1km 내 명소 아이콘 생성

+++ src/main/java/com/project_merge/jigu_travel/api/board/repository/BoardJpaRepository.java
@@ -1,10 +1,11 @@
 package com.project_merge.jigu_travel.api.board.repository;
 
+
 import com.project_merge.jigu_travel.api.board.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardJpaRepository extends JpaRepository<Board, Integer> {
+public interface BoardJpaRepository extends JpaRepository<Board, Long> {
 
 }
